# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Secure Online Election Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user status or role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── ELECTION REQUESTS ─────────────────────────────────────────────────────
...
  /elections/{electionId}/voters:
    get:
      operationId: getElectionVoters
      tags: [voters]
      summary: List voters for an election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voter"
    post:
      operationId: joinElection
      tags: [voters]
      summary: Join an election (voter)
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voter"
  /elections/{electionId}/voters/{voterId}/approve:
    post:
      operationId: approveVoter
      tags: [voters]
      summary: Approve a voter
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
        - name: voterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voter"

  # ─── SECRET IDs ────────────────────────────────────────────────────────────
  /elections/{electionId}/secret-ids:
    get:
      operationId: listSecretIds
      tags: [voters]
      summary: List secret IDs for election
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Secret IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecretId"
    post:
      operationId: generateSecretIds
      tags: [voters]
      summary: Generate secret IDs for all finalized voters
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── VOTES ─────────────────────────────────────────────────────────────────
  /elections/{electionId}/vote:
    post:
      operationId: castVote
      tags: [votes]
      summary: Cast a vote
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteInput"
      responses:
        "201":
          description: Vote cast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Already voted or invalid secret ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /elections/{electionId}/my-vote:
    get:
      operationId: getMyVote
      tags: [votes]
      summary: Check if current user has voted
      parameters:
        - name: electionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vote status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteStatus"

  # ─── RESULTS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, fullName, email, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, election_creator, voter]
        status:
          type: string
          enum: [active, blocked, pending]
        orgName:
          type: ["string", "null"]
        electionPurpose:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, blocked]
        role:
          type: string
          enum: [super_admin, election_creator, voter]
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── ELECTION REQUESTS ───────────────────────────────────────────────────
...
    Voter:
      type: object
      required: [id, electionId, userId, status, joinedAt]
      properties:
        id:
          type: integer
        electionId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, waitlisted, finalized]
        hasVoted:
          type: boolean
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    SecretId:
      type: object
      required: [id, voterId, electionId, maskedSecretId, isUsed, createdAt]
      properties:
        id:
          type: integer
        voterId:
          type: integer
        electionId:
          type: integer
        maskedSecretId:
          type: string
        isUsed:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── VOTES ───────────────────────────────────────────────────────────────
    VoteInput:
      type: object
      required: [candidateId, secretId]
      properties:
        candidateId:
          type: integer
        secretId:
          type: string
    VoteStatus:
      type: object
      required: [hasVoted]
      properties:
        hasVoted:
          type: boolean
        votedAt:
          type: ["string", "null"]
          format: date-time

    # ─── RESULTS ─────────────────────────────────────────────────────────────
...