# lib/api-zod/src/generated/types/voteStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Secure Online Election Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface VoteStatus {
  hasVoted: boolean;
  /** @nullable */
  votedAt?: Date | null;
}